requirements.txt
Pillow
aggdraw
pyshp
pyproj
numpy

